Extracted regexp matcher to testutils
